requirements.txt
fastapi==0.115.0
pydantic-settings==2.5.2
chromadb==0.5.20
uvicorn==0.30.6
openai==1.51.0
mem0ai==0.1.29
python-telegram-bot==21.5
pypdf==4.3.1
beautifulsoup4==4.12.3
requests==2.32.3
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
markdown==3.7
